HTML: tags and attributes refactoring.
Now all prefixes, namespaces, tags, attribute names are numbers and their real
names are found in the hash table.

Changes:
    node->tag_id rename to node->local_name
    Remove attributes name, local_name. Now this is all transferred to the node.
    local_name to attr->node.local_name, name to attr->qualified_name.

+++ source/lexbor/css/syntax/token.h
@@ -83,7 +83,7 @@ struct lxb_css_syntax_token_data {
     lexbor_in_node_t               *node_done;
     lxb_status_t                   status;
     int                            count;
-    size_t                         num;
+    uint32_t                       num;
     bool                           is_last;
 };
